+++ i18n/locales/en.json
@@ -35,6 +35,7 @@
     "previewHintDescription": "Browse curated products, craft new listings, authenticate with demo accounts, and inspect customer profiles without leaving this experience.",
     "featuredTitle": "Curated picks",
     "featuredSubtitle": "A snapshot of products fetched directly from the Fake Store API.",
+    "featuredEmpty": "No featured products available yet.",
     "actions": {
       "browseProducts": "Browse products",
       "createProduct": "Create product",
