+++ pages/products/new.vue
@@ -53,6 +53,17 @@ const validate = () => {
   return true
 }
 
+const updateFormField = (field: keyof NewProductPayload, value: string | number) => {
+  if (field === 'price') {
+    const numeric = typeof value === 'number' ? value : Number(value)
+    if (Number.isFinite(numeric)) {
+      form.price = numeric
+    }
+    return
+  }
+  form[field] = value as any
+}
+
 const onSubmit = async () => {
   apiError.value = ''
   result.value = null
@@ -72,101 +83,16 @@ const onSubmit = async () => {
 
 <template>
   <section class="mx-auto max-w-4xl px-4 py-12 sm:px-6 lg:px-8">
-    <div class="space-y-4">
-      <p class="text-sm font-semibold uppercase tracking-wide text-primary-600">{{ t('createProduct.badge') }}</p>
-      <h1 class="text-3xl font-bold text-slate-900">{{ t('createProduct.title') }}</h1>
-      <p class="text-slate-600">{{ t('createProduct.subtitle') }}</p>
-    </div>
-
-    <form class="mt-8 space-y-6" @submit.prevent="onSubmit">
-      <div class="grid gap-6 lg:grid-cols-2">
-        <label class="space-y-2">
-          <span class="text-sm font-semibold text-slate-700">{{ t('createProduct.fields.title') }}</span>
-          <input
-            v-model="form.title"
-            type="text"
-            class="w-full rounded-2xl border border-slate-200 px-4 py-2.5 text-sm focus:border-primary-300 focus:outline-none focus:ring-2 focus:ring-primary-200"
-            :placeholder="t('createProduct.placeholders.title')"
-          />
-          <span v-if="errors.title" class="text-xs text-red-500">{{ errors.title }}</span>
-        </label>
-
-        <label class="space-y-2">
-          <span class="text-sm font-semibold text-slate-700">{{ t('createProduct.fields.price') }}</span>
-          <input
-            v-model.number="form.price"
-            type="number"
-            min="1"
-            step="0.01"
-            class="w-full rounded-2xl border border-slate-200 px-4 py-2.5 text-sm focus:border-primary-300 focus:outline-none focus:ring-2 focus:ring-primary-200"
-            :placeholder="t('createProduct.placeholders.price')"
-          />
-          <span v-if="errors.price" class="text-xs text-red-500">{{ errors.price }}</span>
-        </label>
-      </div>
-
-      <label class="space-y-2">
-        <span class="text-sm font-semibold text-slate-700">{{ t('createProduct.fields.image') }}</span>
-        <input
-          v-model="form.image"
-          type="url"
-          class="w-full rounded-2xl border border-slate-200 px-4 py-2.5 text-sm focus:border-primary-300 focus:outline-none focus:ring-2 focus:ring-primary-200"
-          :placeholder="t('createProduct.placeholders.image')"
-        />
-        <span v-if="errors.image" class="text-xs text-red-500">{{ errors.image }}</span>
-      </label>
-
-      <label class="space-y-2">
-        <span class="text-sm font-semibold text-slate-700">{{ t('createProduct.fields.category') }}</span>
-        <select
-          v-model="form.category"
-          class="w-full rounded-2xl border border-slate-200 px-4 py-2 text-sm font-semibold text-slate-600 focus:border-primary-300 focus:outline-none focus:ring-2 focus:ring-primary-200"
-        >
-          <option value="" disabled>{{ t('createProduct.placeholders.category') }}</option>
-          <option v-for="option in categoryOptions" :key="option" :value="option">
-            {{ option }}
-          </option>
-        </select>
-        <span v-if="errors.category" class="text-xs text-red-500">{{ errors.category }}</span>
-      </label>
-
-      <label class="space-y-2">
-        <span class="text-sm font-semibold text-slate-700">{{ t('createProduct.fields.description') }}</span>
-        <textarea
-          v-model="form.description"
-          rows="6"
-          class="w-full rounded-2xl border border-slate-200 px-4 py-2.5 text-sm focus:border-primary-300 focus:outline-none focus:ring-2 focus:ring-primary-200"
-          :placeholder="t('createProduct.placeholders.description')"
-        />
-        <span v-if="errors.description" class="text-xs text-red-500">{{ errors.description }}</span>
-      </label>
-
-      <div v-if="apiError" class="rounded-2xl border border-red-200 bg-red-50 px-4 py-3 text-sm text-red-600">
-        {{ apiError }}
-      </div>
-
-      <div class="flex items-center justify-between">
-        <NuxtLink to="/products" class="text-sm font-semibold text-slate-500 hover:text-slate-700">
-          {{ t('actions.cancel') }}
-        </NuxtLink>
-        <button
-          type="submit"
-          class="link-button px-6 text-base"
-          :disabled="submitting"
-        >
-          <span v-if="submitting" class="flex items-center gap-2">
-            <span class="h-4 w-4 animate-spin rounded-full border-2 border-slate-200 border-t-primary-600" />
-            {{ t('createProduct.actions.creating') }}
-          </span>
-          <span v-else>
-            {{ t('createProduct.actions.submit') }}
-          </span>
-        </button>
-      </div>
-    </form>
-
-    <div v-if="result" class="mt-8 rounded-3xl border border-emerald-200 bg-emerald-50 px-6 py-5 text-sm text-emerald-700">
-      {{ t('createProduct.success', { id: result.id }) }}
-    </div>
+    <ProductCreateHeader />
+    <ProductCreateForm
+      :form="form"
+      :errors="errors"
+      :categories="categoryOptions"
+      :submitting="submitting"
+      :api-error="apiError"
+      @update="updateFormField"
+      @submit="onSubmit"
+    />
+    <ProductCreateSuccess v-if="result" :id="result.id" />
   </section>
 </template>
